requirements.txt
openai
fastapi
uvicorn
python-dotenv
pydantic
sqlalchemy 
pyodbc 
alembic 
chromadb 
langchain 
langchain-community 
langchain-openai 
tiktoken

